Feat: 컬럼 정렬 기능 및 화면 추가

+++ src/main/java/com/example/collaborativetools/column/dto/request/ColumnCreateRequest.java
@@ -18,7 +18,7 @@ public class ColumnCreateRequest {
 	private final String title;
 
 	@NotNull
-	private final Integer sequence;
+	private final Double sequence;
 
 	public Columns toEntity(Board entity) {
 		return Columns.create(title, sequence, entity);
